Update start command to run
When working on the SoapUI executor blog post, it was brought to my attention
that we prefer to use run over start to execute tests. This commit updates
the occurences of start in the docs.
I also noticed that the latest Testkube CLI does not print the ASCII,
but it still appears in the docs, so I removed those as well..