Use a Docker image to encapsulate proto tooling versions
Rather than force installing certain versions of protoc and the related go tooling
that we use, instead we can encapsulate them within a Dockerfile to ensure that
it is easier for contributors to maintain continuity of compiler and plugin versions.

Note: This initial commit does not include the `generate-protobuf` Make rule which
generates the log protobufs as those generated files are already using newer versions
of the go tooling than is used for the service protobufs. This is an example of why
proper encapsulation of compiler and plugin versions is important. We will need to
discuss whether to update the service protobuf tooling versions to match the logs,
or perhaps to update all associated tooling to more recent versions (current versions
are from 2022). At which point it would be quite simple to merge the Makefile rules.