feat: Add Execution CRD control to Agent (#6386)
* Create controllers for execution CRDs

These controllers are implemented differently from the similar CRD watching constructs
already implemented in the agent and we should take time to migrate those other CRD
synchronisation "controllers" to a similar structure.

* Initialise CRD controllers on agent startup

Operation of the controllers is hidden behind the `ENABLE_K8S_CONTROLLERS` flag
as they would otherwise conflict with deployed `testkube-operator` instances.

Controllers for legacy CRDs have been separated to make it clearer how to
remove them in the future.

* Update testkube devbox for new operator deployment

The devbox creates its own deployment of the testkube agent and so those
resources have been updated to include the correct Role permissions and to include
an optional `--enable-k8s-controllers` flag.

* Add tests for workflow controller

Currently skipping unit tests for legacy controllers with the expectation that they
will be removed soon (that could be a mistake).

* Clean up legacy cronjobs

When making use of the new cronjob scheduler, first clean up any legacy CronJob
resources that may exist in order to prevent duplicate runs of scheduled tests.

* Disable controllers for non-superagent agent types