Enable runner self registration in certain circumstances
This functionality allows for runners to be created and registered with the control
plane in a single step by external tools such as via Helm installs.

Some quite serious assumptions are made here and I still need to work out where to
persist the generated runner key otherwise the agent will attempt to re-register
itself on every startup, which may not be ideal.