feat: enable runner self-registration (#6410)
* Enable runner self registration in certain circumstances

This functionality allows for runners to be created and registered with the control
plane in a single step by external tools such as via Helm installs.

Some quite serious assumptions are made here and I still need to work out where to
persist the generated runner key otherwise the agent will attempt to re-register
itself on every startup, which may not be ideal.

* Attempt to use a Kubernetes secret to persist agent keys

This will require permissions to create the secret and an optional envFrom secret
to be added to the helm chart.

* Add a configuration flag to switch between fixed and floating runners

By default we assume that the runner is a Fixed type runner as that is what most
customers are using.

* Use gRPC to register runners

This allows for runner registration without having to use a separate
communication mechanism to normal runner activities.

* Remove unnecessary environment ID from self registration

* Improve self registration validation and error checking

* Use a config value for the runner name

This allows it to be set to a unique value at deployment time.

We do not want to generate a unique value at run time because that
would invalidate any uniqueness validation server-side that attempts
to defend against runaway registration loops.