fix: assignment lock logic (#6482)
The locking via assignment logic was not quite tight.

The runnerid can be a non-nil blank string, because when there is an
error sending start request to the runner the orchestrator immediately
tries to unassign the execution by setting it to blank string. If this
happens, then it seems the execution will then be unassignable through
the current logic. The GetUnassigned logic takes this case into account
returning the values with blank strings as well, but it does not seem
Assign is able to handle such a precondition.

Another possible problem is with reassignment, before reassigning an
execution to a different runner, the lock should enforce an assignment
timeout this should help avoid multiple orchestrator instances from
reassigning an execution in short order. Particularly this could be
happening in the recovery loop after an execution has been assigned for
one minute or longer. The one minute timeout in these conditions comes
from the recovery loop period and already logic inside that seems to
attempt to do the same thing without accounting for concurrency
orchestrator instances.

The immediate unassigning on start request failure will need to be
removed also because of the above, and in this cases it should just
fall back to the one minute recovery loop for the sake of simplicity.

Other guard rails put in place to make sure the assigned at time
monotonicly increases.