fix: execution initialisation (#6615)
* During initialisation set executions to SCHEDULING

When the Init RPC is called we know that the runner is starting to
execute the execution and so it can transition to the SCHEDULING
status.

* Allow scheduling to be finished

If an execution has reached a SCHEDULING state it should be allowed
to be finished as this represents the state of failing to be
scheduled in kubernetes and then being aborted or cancelled as a
result.