feat: execution polling (#6617)
* Add new protos for polling executions

Updates the proto generation
configuration to use up to date tool
versions allowing use of the latest
proto editions features.

Generation for older proto definitions
is preserved using a separate
configuration file and exclusion
directives. These can be removed once we
have migrated all proto definitions to
the new structure and generation.

* Add buf github action

This could not be added before because we could
not use the buf tool to validate the existing
protos as they were so out of spec and so at
least one valid proto was required for the tool
to be used.