Reduce the log level of gRPC timeouts
This is normal and expected behaviour of the
system now we are doing away with the ping
pong game that we were playing with the server
(which is an awful abuse of gRPC). However,
these "error" and reconnection messages were
troubling to see in log output, even though
they are expected and normal behaviour now.

So the level has been reduced to INFO, unless
an actual error has occurred in which case they
will still be emitted at the ERROR level.