chore: Reduce the log level of gRPC timeouts (#6702)
* Reduce the log level of gRPC timeouts

This is normal and expected behaviour of the
system now we are doing away with the ping
pong game that we were playing with the server
(which is an awful abuse of gRPC). However,
these "error" and reconnection messages were
troubling to see in log output, even though
they are expected and normal behaviour now.

So the level has been reduced to INFO, unless
an actual error has occurred in which case they
will still be emitted at the ERROR level.

* Reconnect gRPC streams individually

Rather than having streams killed if one of them
needs to reconnect, instead streams should be
individually retried and reconnected so that
a single failure does not kill all of them.