fix: event loop unimplemented backoff
In the event that the Control Plane does not
support the event stream endpoint, we should not
exit and cause a reconnection attempt.

Instead we should backoff and try again at a
later time. The time to backoff is set to start
at a 1 day growing to a maximum of 7 days.