feat: Add RPC definitions for sync functionality (#6756)
Whilst I hate that these RPCs use `bytes` to
carry structured data, my attempts to generate
or manually copy these data structures into
proto definitions were fruitless so until we
have some sort of CRD generation that can
create protobuf definitions for us this is the
best we can hope for. I did try `go-to-protobuf`
which is what the Kubernetes project uses, and
whilst it was mostly fine it would have required
some wider changes to the repository to make it
work nicely with the other protobuf changes that
we have here, so maybe in the future for a v2
of these RPCs we can implement them, but for
now this is the best we can hope for.