fix: avoid warning of insecure grpc connections from execution pods (#6775)
Yes the variable name is long; no I do not think
we should change it. See the comment for that
variable. People don't read docs, and shorter
variable names still get missed. TLS is "hard"
but that is not really a good excuse for just
not doing any security at all.