feat: enable new sync agent (#6797)
* feat: Implement sync Kubernetes controllers

These controllers listen for changes on objects
that we wish to sync and then update an
abstract store interface depending on the
observed state of the object.

That abstract store interface is implemented by
the grpc client so these can all be neatly
aligned with the previous work.

* chore: remove IsSupported

For sync we can use the Control Plane capability
to validate whether or not support exists on the
Control Plane for the feature and so having a
separate function to check this is not required.

* chore: export store interfaces so that they can be used later to compose other interfaces

* feat: create a dummy implementation of a store

This can be used as a fallback for when the main
store cannot be used for some reason (missing
support on the Control Plane) without having
to totally avoid startup of controllers.

* feat: start sync controllers when flag is enabled

* chore: remove no longer used sync flags and code

* fix: new RPC sync

Correctly apply API token for new
RPC sync connections.

* fix: helm charts

Remove unused gitops/sync values from
helm charts.