fix: cronjob logic
After the refactor of the cronjob system a glaring issue with the way that
object deletion has been handled became apparent. This likely has been a
cause of ballooning schedule counts in Kubernetes, as well as potentially
causing duplicate schedule firing, or firing schedules for no longer
existent workflows.

To combat this the way in which changes are communicated is modified to
correctly allow for removals to be implemented, by their omission. This
requires that the scheduling manager must replace all existing workflow
schedules with only the schedules that are passed to it.

This has resulted in a renaming of the manager interface and changes to
the other defined interfaces to facilitate the passing of multiple schedules
in a single message rather than as individual messages.

More information is also provided for workflow identification to
permit multi-tenancy use cases.