chore: Remove parallel test markers from unit tests
Cleaned up parallel test markers due to no benefit of them in most cases
when it comes to unit tests.

Parallel flag for tests is a really good thing when you have a
long running process that can be nicely parallelized and even then the
benefit is only isolated to packages:
1. go into package
2. iterate over tests
3. execute all of them and pause parallel tests
4. wait for non parallel to finish
5. run all parallel tests

For very fast tests, this is in most cases just not worth it and the
overhead (although non existent) just eats all the gains and can even be
net negative.

Running tests both with and without `Parallel()` I did not notice any degradation in
test run speeds.

Signed-off-by: afterlook <29754364+afterlook@users.noreply.github.com>