feat: silence unnecessary executions state transitions
Sometimes the Control Plane does not know that an execution has been aborted or cancelled. This can lead to the Control Plane requesting over and over to perform a state transition on these executions and the Agent logging that the attempted transition failed because the execution no longer exists.

This changes those specific messages to be only emitted at the debug level, whilst leaving the error logging for other errors that may occur.