fix: include timezone data in application
Not all hosts will have the tzdb configured,
this provides a fallback but does increase
binary size quite significantly.

It _should_ only be imported by the `main`
package and was not previously, as a result
it was erroneously removed.

A better way to ensure this cannot happen again
would be to add the `-tags timetzdata` flag to
build arguments as then it cannot be removed
by code changes.