fix: include timezone data in application (#6920)
Not all hosts will have the tzdb configured,
this provides a fallback but does increase
binary size quite significantly.

It _should_ only be imported by the `main`
package and was not previously, as a result
it was erroneously removed.

A better way to ensure this cannot happen again
would be to add the `-tags timetzdata` flag to
build arguments as then it cannot be removed
by code changes.

(cherry picked from commit ff0220f6fc7af230e0b1a196eb8002f69fe2e297)